+++ src/main/java/org/bigml/binding/resources/AbstractResource.java
@@ -54,6 +54,7 @@ public abstract class AbstractResource {
     public final static String EXECUTION_PATH = "execution";
     public final static String LIBRARY_PATH = "library";
     public final static String ASSOCIATION_PATH = "association";
+    public final static String ASSOCIATIONSET_PATH = "associationset";
 
 
     // Base Resource regular expressions
@@ -94,6 +95,8 @@ public abstract class AbstractResource {
     static String EXECUTION_RE = "^" + EXECUTION_PATH + "/[a-f,0-9]{24}$";
     static String LIBRARY_RE = "^" + LIBRARY_PATH + "/[a-f,0-9]{24}$";
     static String ASSOCIATION_RE = "^" + ASSOCIATION_PATH
+            + "/[a-f,0-9]{24}$";
+    static String ASSOCIATIONSET_RE = "^" + ASSOCIATIONSET_PATH
             + "/[a-f,0-9]{24}$";
 
     // HTTP Status Codes from https://bigml.com/developers/status_codes
@@ -172,6 +175,7 @@ public abstract class AbstractResource {
     protected String EXECUTION_URL;
     protected String LIBRARY_URL;
     protected String ASSOCIATION_URL;
+    protected String ASSOCIATIONSET_URL;
 
     public final static String DOWNLOAD_DIR = "/download";
 
@@ -202,6 +206,7 @@ public abstract class AbstractResource {
             EXECUTION_URL = BIGML_URL + EXECUTION_PATH;
             LIBRARY_URL = BIGML_URL + LIBRARY_PATH;
             ASSOCIATION_URL = BIGML_URL + ASSOCIATION_PATH;
+            ASSOCIATIONSET_URL = BIGML_URL + ASSOCIATIONSET_PATH;
 
             this.cacheManager = cacheManager;
         } catch (AuthenticationException ae) {